# kernel-openapi.yaml
openapi: 3.1.0
info:
  title: Institutional Kernel API
  version: "1.1.0"
  description: >
    API HTTP del Institutional Kernel de Mi Rotaract. Fuente única de verdad
    para identidad, organizaciones jerárquicas, membresías, períodos, cargos
    y autoridades, autorización contextual, solicitudes de ingreso,
    transferencias entre organizaciones y registro de módulos externos.
    Este documento es un contrato de datos derivado de `kernel-spec.md`
    (secciones 4, 5, 9 y 10). No reemplaza la especificación; ante cualquier
    conflicto, `kernel-spec.md` es la fuente normativa. El proceso de
    compatibilidad y publicación está definido en
    `kernel-openapi-contract-plan.md`.
  contact:
    name: Institutional Kernel
  license:
    name: Uso interno — Mi Rotaract
    identifier: LicenseRef-MiRotaract-Internal
...
servers:
  - url: https://api.agendai.com.ar/api/kernel/v1
    description: Producción
  - url: https://staging-api.agendai.com.ar/api/kernel/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /membership-applications:
    post:
      operationId: createMembershipApplication
      tags: [Applications]
      summary: Crear una solicitud de ingreso (DRAFT)
      x-required-permission: kernel.application.create.self
      requestBody:
        required: true
        content:
          application/json:
            schema:
              {
                $ref: "#/components/schemas/CreateMembershipApplicationRequest",
              }
      responses:
        "201":
          description: Creada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipApplication" }
        "409":
          description: Ya existe una solicitud abierta (CA-SOL-01).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      operationId: listMembershipApplications
      tags: [Applications]
      summary: Listar solicitudes de ingreso
      description: >
        Requiere `kernel.application.review` para ver solicitudes de
        terceros; sin ese permiso, `kernel.application.read.self` limita el
        resultado a las propias y el filtro `personId` se ignora (se fuerza
        al actor).
      x-required-permission: kernel.application.review
      parameters:
        - name: organizationId
          in: query
          schema: { type: string }
        - name: personId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ApplicationStatus" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MembershipApplication" }
  /membership-applications/{applicationId}:
    get:
      operationId: getMembershipApplication
      tags: [Applications]
      summary: Obtener una solicitud
      x-required-permission: kernel.application.read.self
      parameters:
        - $ref: "#/components/parameters/applicationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipApplication" }
        "404": { $ref: "#/components/responses/NotFound" }
  /membership-applications/{applicationId}/submit:
    post:
      operationId: submitMembershipApplication
      tags: [Applications]
      summary: Enviar una solicitud (DRAFT → SUBMITTED)
      x-required-permission: kernel.application.create.self
      parameters:
        - $ref: "#/components/parameters/applicationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipApplication" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /membership-applications/{applicationId}/approve:
    post:
      operationId: approveMembershipApplication
      tags: [Applications]
      summary: Aprobar una solicitud (SUBMITTED → APPROVED)
      description: >
        Crea o reactiva la membresía en la misma transacción
        (invariante 6.8.3-4, CA-SOL-02).
      x-required-permission: kernel.application.review
      parameters:
        - $ref: "#/components/parameters/applicationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipApplication" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /membership-applications/{applicationId}/reject:
    post:
      operationId: rejectMembershipApplication
      tags: [Applications]
      summary: Rechazar una solicitud (SUBMITTED → REJECTED)
      description: Exige motivo (invariante 6.8.5, CA-SOL-03).
      x-required-permission: kernel.application.review
      parameters:
        - $ref: "#/components/parameters/applicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rejectionReason]
              properties:
                rejectionReason: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipApplication" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /membership-applications/{applicationId}/cancel:
    post:
      operationId: cancelMembershipApplication
      tags: [Applications]
      summary: Cancelar una solicitud propia (DRAFT/SUBMITTED → CANCELLED)
      description: Sólo se cancela desde `DRAFT` o `SUBMITTED` (invariante 6.8.6).
      x-required-permission: kernel.application.cancel.self
      parameters:
        - $ref: "#/components/parameters/applicationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipApplication" }
        "409": { $ref: "#/components/responses/InvalidTransition" }

  # ---------------------------------------------------------------------
  # 9.10 Transferencias
  # ---------------------------------------------------------------------
  /membership-transfers:
    post:
      operationId: requestMembershipTransfer
      tags: [Transfers]
      summary: Solicitar una transferencia de membresía (REQUESTED)
      description: >
        Sólo se transfiere una membresía `ACTIVE` (invariante 6.9.1,
        CA-TRA-01). Origen y destino deben ser diferentes
        (invariante 6.9.2). Sólo una transferencia abierta por membresía
        (invariante 6.9.3, CA-TRA-06).
      x-required-permission: kernel.transfer.create.self
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: "#/components/schemas/RequestMembershipTransferRequest" }
      responses:
        "201":
          description: Creada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipTransfer" }
        "409":
          description: Ya existe una transferencia abierta para esa membresía.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      operationId: listMembershipTransfers
      tags: [Transfers]
      summary: Listar transferencias
      description: >
        Sin `kernel.transfer.accept`/`.confirm`/`.reject` (alcance staff), el
        resultado se limita a las transferencias solicitadas por el propio
        actor (`kernel.transfer.read.self`); el filtro `requestedById` se
        ignora en ese caso y se fuerza al actor.
      x-required-permission: kernel.transfer.read.self
      parameters:
        - name: membershipId
          in: query
          schema: { type: string }
        - name: fromOrganizationId
          in: query
          schema: { type: string }
        - name: toOrganizationId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/TransferStatus" }
        - name: requestedById
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MembershipTransfer" }
  /membership-transfers/{transferId}:
    get:
      operationId: getMembershipTransfer
      tags: [Transfers]
      summary: Obtener una transferencia
      x-required-permission: kernel.transfer.read.self
      parameters:
        - $ref: "#/components/parameters/transferId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipTransfer" }
        "404": { $ref: "#/components/responses/NotFound" }
  /membership-transfers/{transferId}/accept:
    post:
      operationId: acceptTransferByDestination
      tags: [Transfers]
      summary: Aceptar desde la organización destino (REQUESTED → ACCEPTED_BY_DESTINATION)
      description: El destino acepta antes de que el origen confirme (invariante 6.9.4, CA-TRA-02).
      x-required-permission: kernel.transfer.accept
      parameters:
        - $ref: "#/components/parameters/transferId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipTransfer" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /membership-transfers/{transferId}/confirm:
    post:
      operationId: confirmTransferByOrigin
      tags: [Transfers]
      summary: Confirmar desde la organización origen (ACCEPTED_BY_DESTINATION → CONFIRMED_BY_ORIGIN)
      description: CA-TRA-03. Puede disparar `complete` automáticamente según configuración (§9.10).
      x-required-permission: kernel.transfer.confirm
      parameters:
        - $ref: "#/components/parameters/transferId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipTransfer" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /membership-transfers/{transferId}/complete:
    post:
      operationId: completeMembershipTransfer
      tags: [Transfers]
      summary: Completar la transferencia (CONFIRMED_BY_ORIGIN → COMPLETED)
      description: >
        Operación administrativa/interna o automática tras `confirm` (§9.10).
        En una única transacción: marca la membresía origen `TRANSFERRED`,
        finaliza cargos activos incompatibles, crea o reactiva la membresía
        destino, guarda `destinationMembershipId` y emite un único evento
        (invariante 6.9.5-6, CA-TRA-04, CA-TRA-05). Es inmutable una vez
        completada (invariante 6.9.8).
        (Supuesto) Se asume que requiere `kernel.transfer.confirm` cuando se
        invoca manualmente; §10.1 no define un permiso "complete" propio.
      x-required-permission: kernel.transfer.confirm
      parameters:
        - $ref: "#/components/parameters/transferId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipTransfer" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /membership-transfers/{transferId}/reject:
    post:
      operationId: rejectMembershipTransfer
      tags: [Transfers]
      summary: Rechazar una transferencia
      description: Exige motivo (invariante 6.9.7).
      x-required-permission: kernel.transfer.reject
      parameters:
        - $ref: "#/components/parameters/transferId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rejectionReason]
              properties:
                rejectionReason: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipTransfer" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /membership-transfers/{transferId}/cancel:
    post:
      operationId: cancelMembershipTransfer
      tags: [Transfers]
      summary: Cancelar una transferencia propia
      description: >
        (Supuesto) Se asume `kernel.transfer.create.self` como permiso
        requerido, por ser el solicitante quien cancela su propia
        solicitud; §10.1 no define un permiso "cancel" dedicado.
      x-required-permission: kernel.transfer.create.self
      parameters:
        - $ref: "#/components/parameters/transferId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipTransfer" }
        "409": { $ref: "#/components/responses/InvalidTransition" }

  # ---------------------------------------------------------------------
  # 9.11 Módulos
  # ---------------------------------------------------------------------
  /modules:
    post:
      operationId: registerModule
      tags: [Modules]
      summary: Registrar un módulo
      description: >
        Un módulo se registra sin migrar tablas del kernel (CA-MOD-01). El
        registro administra metadatos; no ejecuta código externo (decisión
        final #10).
      x-required-permission: kernel.module.register
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateModuleRequest" }
      responses:
        "201":
          description: Creado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModuleDefinition" }
    get:
      operationId: listModules
      tags: [Modules]
      summary: Listar módulos registrados
      x-required-permission: kernel.module.read
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ModuleStatus" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ModuleDefinition" }
  /modules/{moduleId}:
    get:
      operationId: getModule
      tags: [Modules]
      summary: Obtener un módulo
      x-required-permission: kernel.module.read
      parameters:
        - $ref: "#/components/parameters/moduleId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModuleDefinition" }
        "404": { $ref: "#/components/responses/NotFound" }
  /modules/{moduleId}/manifest:
    put:
      operationId: updateModuleManifest
      tags: [Modules]
      summary: Reemplazar el manifiesto de un módulo
      x-required-permission: kernel.module.register
      parameters:
        - $ref: "#/components/parameters/moduleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifest]
              properties:
                manifest: { type: object }
                configurationSchema: { type: object, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModuleDefinition" }
  /modules/{moduleId}/deprecate:
    post:
      operationId: deprecateModule
      tags: [Modules]
      summary: Marcar un módulo como DEPRECATED
      description: Un módulo `DEPRECATED` no admite instalaciones nuevas (invariante 6.10.3).
      x-required-permission: kernel.module.register
      parameters:
        - $ref: "#/components/parameters/moduleId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModuleDefinition" }
  /organizations/{organizationId}/modules/{moduleId}/install:
    post:
      operationId: installModule
      tags: [Modules]
      summary: Instalar un módulo en una organización (PENDING)
      description: Una organización tiene como máximo una instalación por módulo (invariante 6.10.2).
      x-required-permission: kernel.module.install
      parameters:
        - $ref: "#/components/parameters/organizationId"
        - $ref: "#/components/parameters/moduleId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                configuration: { type: object, nullable: true }
      responses:
        "201":
          description: Instalada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModuleInstallation" }
        "409":
          description: Ya existe una instalación para ese módulo/organización.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Error" }
  /organizations/{organizationId}/modules/{moduleId}/activate:
    post:
      operationId: activateModuleInstallation
      tags: [Modules]
      summary: Activar una instalación (PENDING/SUSPENDED → ACTIVE)
      description: Valida `configurationSchema` (invariante 6.10.4, CA-MOD-02). Emite evento (CA-MOD-03).
      x-required-permission: kernel.module.install
      parameters:
        - $ref: "#/components/parameters/organizationId"
        - $ref: "#/components/parameters/moduleId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModuleInstallation" }
        "422":
          description: Configuración inválida contra `configurationSchema` (CA-MOD-02).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Error" }
  /organizations/{organizationId}/modules/{moduleId}/configuration:
    patch:
      operationId: updateModuleConfiguration
      tags: [Modules]
      summary: Actualizar la configuración de una instalación
      x-required-permission: kernel.module.configure
      parameters:
        - $ref: "#/components/parameters/organizationId"
        - $ref: "#/components/parameters/moduleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [configuration]
              properties:
                configuration: { type: object }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModuleInstallation" }
        "422":
          description: Configuración inválida contra `configurationSchema`.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Error" }
  /organizations/{organizationId}/modules/{moduleId}/suspend:
    post:
      operationId: suspendModuleInstallation
      tags: [Modules]
      summary: Suspender una instalación (ACTIVE → SUSPENDED)
      x-required-permission: kernel.module.configure
      parameters:
        - $ref: "#/components/parameters/organizationId"
        - $ref: "#/components/parameters/moduleId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModuleInstallation" }
  /organizations/{organizationId}/modules/{moduleId}/disable:
    post:
      operationId: disableModuleInstallation
      tags: [Modules]
      summary: Deshabilitar una instalación
      description: No elimina datos del servicio consumidor (invariante 6.10.5, CA-MOD-04).
      x-required-permission: kernel.module.disable
      parameters:
        - $ref: "#/components/parameters/organizationId"
        - $ref: "#/components/parameters/moduleId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModuleInstallation" }
  /organizations/{organizationId}/modules:
    get:
      operationId: listOrganizationModules
      tags: [Modules]
      summary: Listar instalaciones de módulos de una organización
      x-required-permission: kernel.module.read
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ModuleInstallation" }
  /organizations/{organizationId}/capabilities:
    get:
      operationId: getOrganizationCapabilities
      tags: [Modules]
      summary: Obtener capacidades habilitadas (módulos activos + permisos)
      description: Refleja instalación y permisos (CA-MOD-05).
      x-required-permission: kernel.module.read
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizationId: { type: string }
                  modules:
                    type: array
                    items:
                      type: object
                      properties:
                        moduleId: { type: string }
                        status:
                          { $ref: "#/components/schemas/InstallationStatus" }

  # ---------------------------------------------------------------------
  # 9.12 APIs para servicios
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token de usuario emitido por `POST /auth/login` o
        `/auth/refresh` (§14.2). Vida útil de 10 minutos. `aud`:
        `mirotaract-platform`.
    serviceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token de servicio a servicio firmado (o mTLS equivalente, §14.3).
        `sub` identifica al servicio emisor. `aud` debe ser
        `institutional-kernel`. Requerido para todos los endpoints bajo
        `/service/*`; el kernel rechaza estas rutas con token de usuario.
  parameters:
...
    organizationId:
      name: organizationId
      in: path
      required: true
      schema: { type: string }
...
    moduleId:
      name: moduleId
      in: path
      required: true
      schema: { type: string }
    applicationId:
      name: applicationId
      in: path
      required: true
      schema: { type: string }
    transferId:
      name: transferId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    NotFound:
      description: Recurso inexistente.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidTransition:
      description: La transición de estado solicitada no es válida desde el estado actual.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      description: Problem Details (RFC 9457).
      required: [type, title, status, code, instance]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: Código estable de error, p. ej. `KERNEL_INVALID_TRANSITION`.
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
      example:
        type: https://api.agendai.com.ar/errors/invalid-transition
        title: Invalid state transition
        status: 409
        code: KERNEL_INVALID_TRANSITION
        detail: ACTIVE membership cannot transition directly to PENDING
        instance: /api/kernel/v1/memberships/mem_123/status
        traceId: trc_123
...
    ApplicationStatus:
      type: string
      enum: [DRAFT, SUBMITTED, APPROVED, REJECTED, CANCELLED, EXPIRED]
    TransferStatus:
      type: string
      enum:
        - REQUESTED
        - ACCEPTED_BY_DESTINATION
        - CONFIRMED_BY_ORIGIN
        - COMPLETED
        - REJECTED
        - CANCELLED
        - EXPIRED
...
    ModuleStatus:
      type: string
      enum: [DRAFT, ACTIVE, DEPRECATED, DISABLED]
    InstallationStatus:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED, DISABLED]
...
    MembershipApplication:
      type: object
      required:
        [id, organizationId, requesterPersonId, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        organizationId: { type: string }
        requesterPersonId: { type: string }
        membershipId: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ApplicationStatus" }
        message: { type: string, nullable: true }
        submittedAt: { type: string, format: date-time, nullable: true }
        reviewedById: { type: string, nullable: true }
        reviewedAt: { type: string, format: date-time, nullable: true }
        rejectionReason: { type: string, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateMembershipApplicationRequest:
      type: object
      required: [organizationId]
      properties:
        organizationId: { type: string }
        message: { type: string, nullable: true }
    MembershipTransfer:
      type: object
      required:
        [
          id,
          membershipId,
          fromOrganizationId,
          toOrganizationId,
          requestedById,
          status,
          requestedAt,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: string }
        membershipId: { type: string }
        fromOrganizationId: { type: string }
        toOrganizationId: { type: string }
        requestedById: { type: string }
        status: { $ref: "#/components/schemas/TransferStatus" }
        reason: { type: string, nullable: true }
        requestedAt: { type: string, format: date-time }
        acceptedById: { type: string, nullable: true }
        acceptedAt: { type: string, format: date-time, nullable: true }
        confirmedById: { type: string, nullable: true }
        confirmedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        rejectedById: { type: string, nullable: true }
        rejectedAt: { type: string, format: date-time, nullable: true }
        rejectionReason: { type: string, nullable: true }
        cancelledById: { type: string, nullable: true }
        cancelledAt: { type: string, format: date-time, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        destinationMembershipId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RequestMembershipTransferRequest:
      type: object
      required: [membershipId, toOrganizationId]
      properties:
        membershipId: { type: string }
        toOrganizationId: { type: string }
        reason: { type: string, nullable: true }
...
    CreateModuleRequest:
      type: object
      required: [id, name, version, manifest]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        version: { type: string }
        manifest: { type: object }
        configurationSchema: { type: object, nullable: true }
...